+++ DependencyInjection/Compiler/ApiErrorCodeGuesserCompiler.php
@@ -5,22 +5,22 @@ declare(strict_types=1);
 /*
  * This file is part of the ApiPlatformBundle package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Bundle\ApiPlatformBundle\DependencyInjection\Compiler;
+namespace MarfaTech\Bundle\ApiPlatformBundle\DependencyInjection\Compiler;
 
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
-use Wakeapp\Bundle\ApiPlatformBundle\EventListener\ApiResponseListener;
-use Wakeapp\Bundle\ApiPlatformBundle\Guesser\ApiErrorCodeGuesser;
-use Wakeapp\Bundle\ApiPlatformBundle\Guesser\ApiErrorCodeGuesserInterface;
+use MarfaTech\Bundle\ApiPlatformBundle\EventListener\ApiResponseListener;
+use MarfaTech\Bundle\ApiPlatformBundle\Guesser\ApiErrorCodeGuesser;
+use MarfaTech\Bundle\ApiPlatformBundle\Guesser\ApiErrorCodeGuesserInterface;
 use function is_subclass_of;
 use function sprintf;
 
@@ -31,8 +31,8 @@ class ApiErrorCodeGuesserCompiler implements CompilerPassInterface
      */
     public function process(ContainerBuilder $container): void
     {
-        $guesserId = $container->getParameter('wakeapp_api_platform.error_code_guesser_service');
-        $container->getParameterBag()->remove('wakeapp_api_platform.error_code_guesser_service');
+        $guesserId = $container->getParameter('marfatech_api_platform.error_code_guesser_service');
+        $container->getParameterBag()->remove('marfatech_api_platform.error_code_guesser_service');
 
         if (empty($guesserId)) {
             $guesserId = ApiErrorCodeGuesser::class;