+++ DependencyInjection/Compiler/ApiKernelCompiler.php
@@ -5,20 +5,20 @@ declare(strict_types=1);
 /*
  * This file is part of the ApiPlatformBundle package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Bundle\ApiPlatformBundle\DependencyInjection\Compiler;
+namespace MarfaTech\Bundle\ApiPlatformBundle\DependencyInjection\Compiler;
 
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
-use Wakeapp\Bundle\ApiPlatformBundle\Guesser\ApiAreaGuesserInterface;
-use Wakeapp\Bundle\ApiPlatformBundle\HttpFoundation\ApiKernel;
+use MarfaTech\Bundle\ApiPlatformBundle\Guesser\ApiAreaGuesserInterface;
+use MarfaTech\Bundle\ApiPlatformBundle\HttpFoundation\ApiKernel;
 use function is_subclass_of;
 use function sprintf;
 
@@ -29,8 +29,8 @@ class ApiKernelCompiler implements CompilerPassInterface
      */
     public function process(ContainerBuilder $container): void
     {
-        $apiAreaId = $container->getParameter('wakeapp_api_platform.api_area_guesser_service');
-        $container->getParameterBag()->remove('wakeapp_api_platform.api_area_guesser_service');
+        $apiAreaId = $container->getParameter('marfatech_api_platform.api_area_guesser_service');
+        $container->getParameterBag()->remove('marfatech_api_platform.api_area_guesser_service');
 
         if (!$container->hasDefinition($apiAreaId)) {
             throw new ServiceNotFoundException($apiAreaId, ApiKernel::class);