+++ DependencyInjection/Compiler/ApiResponseListenerCompiler.php
@@ -5,17 +5,17 @@ declare(strict_types=1);
 /*
  * This file is part of the ApiPlatformBundle package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Bundle\ApiPlatformBundle\DependencyInjection\Compiler;
+namespace MarfaTech\Bundle\ApiPlatformBundle\DependencyInjection\Compiler;
 
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Wakeapp\Bundle\ApiPlatformBundle\EventListener\ApiResponseListener;
+use MarfaTech\Bundle\ApiPlatformBundle\EventListener\ApiResponseListener;
 
 class ApiResponseListenerCompiler implements CompilerPassInterface
 {
@@ -24,11 +24,11 @@ class ApiResponseListenerCompiler implements CompilerPassInterface
      */
     public function process(ContainerBuilder $container): void
     {
-        $responseDebug = $container->getParameter('wakeapp_api_platform.response_debug');
-        $container->getParameterBag()->remove('wakeapp_api_platform.response_debug');
+        $responseDebug = $container->getParameter('marfatech_api_platform.response_debug');
+        $container->getParameterBag()->remove('marfatech_api_platform.response_debug');
 
-        $apiResultDtoClass = $container->getParameter('wakeapp_api_platform.api_result_dto_class');
-        $container->getParameterBag()->remove('wakeapp_api_platform.api_result_dto_class');
+        $apiResultDtoClass = $container->getParameter('marfatech_api_platform.api_result_dto_class');
+        $container->getParameterBag()->remove('marfatech_api_platform.api_result_dto_class');
 
         $listenerDefinition = $container->getDefinition(ApiResponseListener::class);
         $listenerDefinition