+++ Dto/ApiResultDtoInterface.php
@@ -5,15 +5,15 @@ declare(strict_types=1);
 /*
  * This file is part of the ApiPlatformBundle package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Bundle\ApiPlatformBundle\Dto;
+namespace MarfaTech\Bundle\ApiPlatformBundle\Dto;
 
-use Wakeapp\Component\DtoResolver\Dto\DtoResolverInterface;
+use MarfaTech\Component\DtoResolver\Dto\DtoResolverInterface;
 
 interface ApiResultDtoInterface extends DtoResolverInterface
 {