+++ EventListener/MinimalApiVersionRequestListener.php
@@ -5,17 +5,17 @@ declare(strict_types=1);
 /*
  * This file is part of the ApiPlatformBundle package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Bundle\ApiPlatformBundle\EventListener;
+namespace MarfaTech\Bundle\ApiPlatformBundle\EventListener;
 
 use Symfony\Component\HttpKernel\Event\RequestEvent;
-use Wakeapp\Bundle\ApiPlatformBundle\Exception\ApiException;
-use Wakeapp\Bundle\ApiPlatformBundle\HttpFoundation\ApiRequest;
+use MarfaTech\Bundle\ApiPlatformBundle\Exception\ApiException;
+use MarfaTech\Bundle\ApiPlatformBundle\HttpFoundation\ApiRequest;
 
 class MinimalApiVersionRequestListener
 {