+++ HttpFoundation/ApiResponse.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 /*
  * This file is part of the ApiPlatformBundle package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Bundle\ApiPlatformBundle\HttpFoundation;
+namespace MarfaTech\Bundle\ApiPlatformBundle\HttpFoundation;
 
 use JsonSerializable;
 use Symfony\Component\HttpFoundation\JsonResponse;