+++ MarfatechApiPlatformBundle.php
@@ -5,21 +5,21 @@ declare(strict_types=1);
 /*
  * This file is part of the ApiPlatformBundle package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Bundle\ApiPlatformBundle;
+namespace MarfaTech\Bundle\ApiPlatformBundle;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
-use Wakeapp\Bundle\ApiPlatformBundle\DependencyInjection\Compiler\ApiErrorCodeGuesserCompiler;
-use Wakeapp\Bundle\ApiPlatformBundle\DependencyInjection\Compiler\ApiKernelCompiler;
-use Wakeapp\Bundle\ApiPlatformBundle\DependencyInjection\Compiler\ApiResponseListenerCompiler;
+use MarfaTech\Bundle\ApiPlatformBundle\DependencyInjection\Compiler\ApiErrorCodeGuesserCompiler;
+use MarfaTech\Bundle\ApiPlatformBundle\DependencyInjection\Compiler\ApiKernelCompiler;
+use MarfaTech\Bundle\ApiPlatformBundle\DependencyInjection\Compiler\ApiResponseListenerCompiler;
 
-class WakeappApiPlatformBundle extends Bundle
+class MarfatechApiPlatformBundle extends Bundle
 {
     /**
      * {@inheritdoc}